fix: create new file override main.qvs
by default it seems even qlik dosent allow empty files there is allways
a default script included if we clear the file and save it, unless we
have 1 line of code.
Create new Files is basicly the same as copy / paste but files are empty
to. So we dont write a file anymore if file content will be empty since we
could not have an empty script.

closes #20