feat: cache connection
- change settings key VsQlik.ConnectionSettings to VsQlik.Servers
- added setting VsCode.Developer

added new setting VsCode.Developer to tell how to cache session cookies, by default this will use only a memory storage which is lost if vscode closes. For Development this is bad since we restart vscode multiple times so all session cookies will stored into a FileStorage.

This is not secure and have to enabled via the settings.json file

```json
VsQlik.Developer: {
  cacheSession: true
}
```