fix: fixed bug multiple authorization requests was running
fixed issue if we added a new workspace folder which requires login credentials, there was multiple authoriaztion calls in hsort time. Which causes the error that we open to many user sessions on qlik sense sever. Now we cache the connections (pending or solved) in authorization helper and return this ones.

closes #273