improve connect / remove from servers
fixed the bug a server which was added as second we could not connect (server not exists, password wrong) was not removed from tree anymore. This happens since we have to wait for the event onDidChangeWorkspacefolders.added to ensure the filesystem has added our server into filetree so we can safely remove it. Otherwise it cant be removed programatically even if vscode.workspace.workspaceFolders tells us the workspace exists, but use vscode.workspace.updateWorkspaceFolders(1, 1) for example will return false.

it also makes it a bit faster to connect now