fix: remove limit for active sessions in qlik sense desktop
for qlik core there is an limit for max open sessions (5) if we dont have a licensed version, this is good enough for development issues but this limit does not exists in qlik sense desktop this reduces the performance leaks on the server. So we just removed the limit for qlik sense desktop but for qlik core it still exists.