fix: secure connections with valid certifcate dont need to be accepted anymore
fixed a bug a server with a secure connection counts as invalid even if there is a valid certificate. it seems this happens because of tls.connect which try to find the server certificate on the computer which fails. Now we run at least 2 server checks for a secure server:

1. server is secure and request does not fail on rejectUnauthorized: true then it is secure and valid
2. server is secure and request does not fail on rejectUnauthorized: false then it is secure but no valid certificate

closes #475