[feat #543] unlock script if document is closed
changed the logic the script will locked as soon it will be opened and unlocked if document is closed again. This will ensure the file is not overridden until it is open, for example editor lost the focus and content will read in again. After document it closed it will unlocked again so next time we get the data from server.